pagerinfo/requirements.txt
flask>=3.0.0
pywebpush>=2.0.0
py-vapid>=1.9.0
pillow>=10.0.0
pyyaml>=6.0
playwright>=1.43.0
